Adopt canary deploy with swarm and traefik

+++ scripts/deployment.groovy
@@ -1,5 +1,5 @@
 // Importante: No final deste ficheiro tens de retornar 'this'
-def deploy(branch) {
+def deploy(branch, strategy) {
 
     def imageTag = ""
 
@@ -18,18 +18,29 @@ def deploy(branch) {
     // Create network
     sh "${ssh} ${branch} 'docker network inspect lms_network >/dev/null 2>&1 || docker network create lms_network'"
 
-    // copy target/LMSBooks-0.0.1-SNAPSHOT.jar
-    sh "scp -o StrictHostKeyChecking=no -F ./deployment-resources/ssh_deployment_config target/LMSBooks-0.0.1-SNAPSHOT.jar ${branch}:/opt/books/${branch}/target/LMSBooks-0.0.1-SNAPSHOT.jar"
+    if(strategy == 'Switch'){
+        // --- INIT SWITCH VERSION
+        sh "scp -o StrictHostKeyChecking=no -F ./deployment-resources/ssh_deployment_config target/LMSBooks-0.0.1-SNAPSHOT.jar ${branch}:/opt/books/${branch}/target/LMSBooks-0.0.1-SNAPSHOT.jar"
+        sh "${ssh} ${branch} 'cd /opt/books/${branch}/ && docker rm -f books_query books_command'"
+        sh "${ssh} ${branch} 'cd /opt/books/${branch}/ && IMAGE_TAG=${imageTag} docker compose pull && docker compose up -d'"
+    } else {
+        // --- INIT CANARY VERSION
+        sh "scp -o StrictHostKeyChecking=no -F ./deployment-resources/ssh_deployment_config target/LMSBooks-0.0.1-SNAPSHOT.jar ${branch}:/opt/books/${branch}/target/LMSBooks-0.0.1-SNAPSHOT.jar"
 
-    // Remove old containers
-    sh "${ssh} ${branch} 'cd /opt/books/${branch}/ && docker rm -f books_query books_command'"
+        // Proxy
+        sh "${ssh} ${branch} 'docker stack deploy -c docker-compose-traefik.yml proxy'"
 
-    // Rollback to tag
-    sh "${ssh} ${branch} 'cd /opt/books/${branch}/ && IMAGE_TAG=${imageTag} docker compose pull && docker compose up -d'"
+        // Stack
+        sh "${ssh} ${branch} 'cd /opt/books/${branch}/ && IMAGE_TAG=${imageTag} docker stack deploy -c docker-compose-swarm.yml ${stackName} --with-registry-auth'"
+    }
 }
 
 def dockerConfig(branch){
+    // Set permissions
     sh 'chmod 600 ./deployment-resources/id_rsa_custom'
+    // Copy rollout files
+    sh "scp -o StrictHostKeyChecking=no -F ./deployment-resources/ssh_deployment_config -r ./rollout ${branch}:/opt/books/${branch}"
+    // Copy docker files
     sh "scp -o StrictHostKeyChecking=no -F ./deployment-resources/ssh_deployment_config *ocker* ${branch}:/opt/books/${branch}"
 }
